repaired some thing

+++ src/main/java/com/RentCar/Database/Clientes.java
@@ -10,7 +10,7 @@ public class Clientes implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long Id;
+    private Integer Id;
     
     private String Nombre;
     
@@ -37,7 +37,7 @@ public class Clientes implements Serializable {
         this.Estado = Estado;
     }
 
-    public Clientes(Long Id, String Nombre, int NoTargetaCR, int LimCredito, String TipoPersona, String Estado) {
+    public Clientes(Integer Id, String Nombre, int NoTargetaCR, int LimCredito, String TipoPersona, String Estado) {
         this.Id = Id;
         this.Nombre = Nombre;
         this.NoTargetaCR = NoTargetaCR;
