+++ src/main/java/com/RentCar/Database/Modelos.java
@@ -11,7 +11,7 @@ public class Modelos implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long Id;
+    private Integer Id;
     
     @OneToOne
     private Marcas marca;
@@ -25,7 +25,7 @@ public class Modelos implements Serializable {
     public Modelos() {
     }    
     
-    public Modelos(Long Id, Marcas marca, String Descripcion, String Estado) {
+    public Modelos(Integer Id, Marcas marca, String Descripcion, String Estado) {
         this.Id = Id;
         this.marca = marca;
         this.Descripcion = Descripcion;
