+++ src/main/java/com/RentCar/Database/Repository/EmpleadosRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface EmpleadosRepo 
-        extends JpaRepository<Empleados, Long> {
+        extends JpaRepository<Empleados, Integer> {
     
 }
